+++ src/controllers/producto.controller.ts
@@ -1,22 +1,53 @@
 import { Request, Response } from 'express';
-
-import { newProductInterface, ProductQueryInterface } from '../interface/producto.inteface';
+import { newProductInterface } from '../interface/producto.inteface';
 import { mongoProductRepository } from '../repositories/mongo';
+import { categoriaRepository } from '../repositories/mongo/categoria.repository';
+import { ProductQueryInterface } from '../interface';
 
 export class ProductoController {
   async getById(req: Request, res: Response) {
     try {
-      let id = req.params.id;
-      let product = await mongoProductRepository.findById(id);
-      res.json(product);
+      let { id } = req.params;
+      if (id) {
+        let product = await mongoProductRepository.findById(id);
+        product ? res.json(product) : res.status(404).json('Product not found');
+      } else {
+        console.log('Entro por el ese');
+        res.status(400).json('Invalid Field: ID');
+      }
     } catch (err) {
       console.log(err);
     }
   }
 
+  async findByCategoria(req: Request, res: Response) {
+    let { categoriaId } = req.params;
+    try {
+      if (categoriaId) {
+        console.log(categoriaId);
+        const productos = await mongoProductRepository.findByCategory(categoriaId);
+        console.log(productos);
+        return res.status(200).json(productos);
+      } else {
+        return res.status(400).json('Invalid params');
+      }
+    } catch (err) {
+      return res.status(500).json(err);
+    }
+  }
+
   async get(req: Request, res: Response) {
     try {
-      let data = await mongoProductRepository.findAll();
+      const { categoria, nombre, minPrice, maxPrice, codigo, minStock, maxStock } = req.query;
+      const query: ProductQueryInterface = {};
+      if (nombre) query.nombre = nombre.toString();
+      if (codigo) query.codigo = Number(codigo);
+      if (minPrice) query.minPrice = Number(minPrice);
+      if (maxPrice) query.maxPrice = Number(maxPrice);
+      if (minStock) query.minStock = Number(minStock);
+      if (maxStock) query.maxStock = Number(maxStock);
+      if (categoria) query.categoria = categoria.toString();
+      let data = await mongoProductRepository.query(query);
       res.json(data);
     } catch (err) {
       console.log(err);
@@ -35,10 +66,17 @@ export class ProductoController {
         stock,
         categoria,
       };
-      let result = await mongoProductRepository.create(producto);
-      return res.status(200).json(result);
+      console.log(categoria);
+      let cat = await categoriaRepository.getCategoriasById(categoria.toString());
+      if (cat) {
+        let result = await mongoProductRepository.create(producto);
+        return res.status(200).json(result);
+      } else {
+        res.status(400).json('Categoria no existente');
+      }
     } catch (err) {
       console.log(err);
+      return res.status(500).json(err);
     }
   }
 
@@ -56,10 +94,13 @@ export class ProductoController {
         categoria,
       };
 
-      if (producto) {
-        let prod = await mongoProductRepository.findById(id);
+      let prod = await mongoProductRepository.findById(id);
+      if (prod) {
         let data = await mongoProductRepository.update(id, producto);
-        res.status(200).json({ producto: 'Producto Actualizado', data });
+        let productToRes = await mongoProductRepository.findById(id);
+        res.status(200).json({ result: 'Producto Actualizado', productToRes });
+      } else {
+        res.status(400).json('Producto no encontrado');
       }
     } catch (err) {
       return res.json(err);
@@ -69,10 +110,15 @@ export class ProductoController {
   async borrar(req: Request, res: Response) {
     try {
       let id = req.params.id;
-      await mongoProductRepository.delete(id);
-      res.status(202).json({
-        msg: 'producto borrado',
-      });
+      let prod = await mongoProductRepository.findById(id);
+      if (prod) {
+        await mongoProductRepository.delete(id);
+        res.status(202).json({
+          msg: 'producto borrado',
+        });
+      } else {
+        res.status(400).json('Producto no encontrado');
+      }
     } catch (err) {
       console.log(err);
     }
